third_party/WebKit/LayoutTests/fast/performance/performance-mark-exceptions.html

<!DOCTYPE html>
<html>
<head>
</head>
<body>
    <script src="../../resources/js-test.js"></script>
    <script>
        description("This tests that 'performance.mark' throws exceptions with reasonable messages.");

        var allTheThings = [
            'navigationStart',
            'unloadEventStart',
            'unloadEventEnd',
            'redirectStart',
            'redirectEnd',
            'fetchStart',
            'domainLookupStart',
            'domainLookupEnd',
            'connectStart',
            'connectEnd',
            'secureConnectionStart',
            'requestStart',
            'responseStart',
            'responseEnd',
            'domLoading',
            'domInteractive',
            'domContentLoadedEventStart',
            'domContentLoadedEventEnd',
            'domComplete',
            'loadEventStart',
            'loadEventEnd',
        ];
        allTheThings.forEach(function(name) {
            shouldThrow('window.performance.mark(\'' + name + '\')', '"SyntaxError: Failed to execute \'mark\' on \'Performance\': \'' + name + '\' is part of the PerformanceTiming interface, and cannot be used as a mark name."');
        });
    </script>
</body>
</html>
