third_party/WebKit/LayoutTests/fast/performance/performance-mark-exceptions-expected.txt

This tests that 'performance.mark' throws exceptions with reasonable messages.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.performance.mark('navigationStart') threw exception SyntaxError: 'navigationStart' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('unloadEventStart') threw exception SyntaxError: 'unloadEventStart' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('unloadEventEnd') threw exception SyntaxError: 'unloadEventEnd' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('redirectStart') threw exception SyntaxError: 'redirectStart' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('redirectEnd') threw exception SyntaxError: 'redirectEnd' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('fetchStart') threw exception SyntaxError: 'fetchStart' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('domainLookupStart') threw exception SyntaxError: 'domainLookupStart' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('domainLookupEnd') threw exception SyntaxError: 'domainLookupEnd' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('connectStart') threw exception SyntaxError: 'connectStart' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('connectEnd') threw exception SyntaxError: 'connectEnd' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('secureConnectionStart') threw exception SyntaxError: 'secureConnectionStart' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('requestStart') threw exception SyntaxError: 'requestStart' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('responseStart') threw exception SyntaxError: 'responseStart' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('responseEnd') threw exception SyntaxError: 'responseEnd' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('domLoading') threw exception SyntaxError: 'domLoading' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('domInteractive') threw exception SyntaxError: 'domInteractive' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('domContentLoadedEventStart') threw exception SyntaxError: 'domContentLoadedEventStart' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('domContentLoadedEventEnd') threw exception SyntaxError: 'domContentLoadedEventEnd' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('domComplete') threw exception SyntaxError: 'domComplete' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('loadEventStart') threw exception SyntaxError: 'loadEventStart' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS window.performance.mark('loadEventEnd') threw exception SyntaxError: 'loadEventEnd' is part of the PerformanceTiming interface, and cannot be used as a mark name..
PASS successfullyParsed is true

TEST COMPLETE
